pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-for-code-agents"
version = "0.1.0"
description = "A multi-agent system that transforms natural language project ideas into structured prompts for code generation"
authors = [
    {name = "Shi Qiu", email = "stephenqius@gmail.com"},
    {name = "Siwei Han", email = "lillian423@126.com"},
    {name = "Zhaorun Chen", email = "billchandler0226@gmail.com"},
    {name = "Huaxiu Yao", email = "huaxiu@cs.unc.edu"},
    {name = "Linjie Li", email = "linjieli222@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core backend framework
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "starlette==0.47.2",
    
    # AI and LLM integration
    "openai==1.101.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "tiktoken==0.9.0",
    "tokenizers==0.21.2",
    
    # File and async operations
    "aiofiles==24.1.0",
    "python-multipart==0.0.20",
    "watchdog==6.0.0",
    
    # Utilities and configuration
    "python-dotenv==1.1.1",
    "python-dateutil==2.8.2",
    "requests==2.32.5",
    "tenacity==8.2.3",
    "jsonschema==4.25.1",
    "regex==2022.1.18",
    "PyYAML==6.0.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "tqdm==4.67.1",
    
    # Google Cloud integrations (optional)
    "google-genai==1.26.0",
    "google-api-python-client==2.176.0",
    "google-auth==2.40.3",
    "google-cloud-aiplatform==1.104.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
agent-for-code-agents = "agent_for_code_agents.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true